fix memory profiler
